pyproject.toml
[project]
name = "pyproj_dep_analyse"
version = "0.0.1"
description = "Python project dependency analyzer"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "rich-click>=1.9.4",
    "lib_cli_exit_tools>=2.1.0",
    "lib_log_rich>=5.3.1",
    "lib_layered_config>=4.0.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
]
license = { text = "MIT" }
authors = [{ name = "bitranox", email = "bitranox@gmail.com" }]
keywords = ["dependency", "analyzer", "pyproject", "outdated", "version", "cli"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Environment :: Console",
  "Operating System :: OS Independent",
  "Typing :: Typed",
]

[project.urls]
Homepage = "https://github.com/bitranox/pyproj_dep_analyse"
Repository = "https://github.com/bitranox/pyproj_dep_analyse.git"
Issues = "https://github.com/bitranox/pyproj_dep_analyse/issues"

[project.optional-dependencies]
dev = [
  "pytest>=9.0.1",
  "pytest-cov>=7.0.0",
  "ruff>=0.14.7",
  "pyright>=1.1.407",
  "bandit>=1.9.2",
  "build>=1.3.0",
  "twine>=6.2.0",
  "codecov-cli>=11.2.5",
  "pip-audit>=2.10.0",
  "textual>=6.7.1",
  "import-linter>=2.7"
]

[project.scripts]
pyproj_dep_analyse = "pyproj_dep_analyse.cli:main"
pyproj-dep-analyse = "pyproj_dep_analyse.cli:main"

[build-system]
requires = ["hatchling>=1.28.0"]
build-backend = "hatchling.build"



[tool.hatch.build.targets.wheel]
packages = ["src/pyproj_dep_analyse"]
include = ["src/pyproj_dep_analyse/py.typed", "src/pyproj_dep_analyse/defaultconfig.toml"]

[tool.ruff]
line-length = 160
target-version = "py313"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
# ignore the "unused import sys" in the jupyter notebook - which we need (it is a linter hickup)
"notebooks/*.ipynb" = ["F401"]

[tool.pyright]
typeCheckingMode = "strict"
extraPaths = [
  "src",
]
exclude = [
  "scripts/menu.py",
]

[tool.pytest.ini_options]
addopts = ["--doctest-modules"]
doctest_optionflags = ["ELLIPSIS", "NORMALIZE_WHITESPACE"]
pythonpath = ["src"]
markers = [
  "os_agnostic: test runs on every supported operating system",
  "os_windows: test exercises Windows-only behavior",
  "os_macos: test exercises macOS-only behavior",
  "os_posix: test exercises behavior requiring POSIX semantics",
  "os_linux: test exercises Linux-only behavior",
]

[tool.coverage.run]
branch = true
parallel = false
data_file = ".coverage.unit"
source = ["src/pyproj_dep_analyse"]

[tool.coverage.report]
omit = [
  "tests/*",
]
fail_under = 85
show_missing = true

[tool.coverage.json]
output = "coverage.json"
pretty_print = true

[tool.clean]
# Patterns to remove when running `make clean` or `python -m scripts clean`
patterns = [
  ".hypothesis",
  ".import_linter_cache",
  ".pytest_cache",
  ".ruff_cache",
  ".pyright",
  ".mypy_cache",
  ".tox",
  ".nox",
  ".eggs",
  "*.egg-info",
  "build",
  "dist",
  "htmlcov",
  ".coverage",
  "coverage.xml",
  "codecov.sh",
  ".cache",
  "result",
]

[tool.git]
# Default git remote for push/release commands
default-remote = "origin"


[tool.scripts.test]
# Configuration for scripts/test.py test runner
pytest-verbosity = "-vv"
coverage-report-file = "coverage.xml"
src-path = "src"
[tool.importlinter]
root_package = "pyproj_dep_analyse"

[[tool.importlinter.contracts]]
name = "Core modules are independent"
type = "independence"
modules = [
  "pyproj_dep_analyse.models",
  "pyproj_dep_analyse.python_version_parser",
]

[[tool.importlinter.contracts]]
name = "CLI is the top layer"
type = "layers"
layers = [
  "pyproj_dep_analyse.cli",
  "pyproj_dep_analyse.analyzer",
  "pyproj_dep_analyse.version_resolver",
  "pyproj_dep_analyse.dependency_extractor",
  "pyproj_dep_analyse.models",
]

[tool.pip-audit]
# Known vulnerabilities to ignore (unfixable or accepted risk)
# GHSA-4xh5-x5gv-qwph: py library ReDoS (no fix available, low risk)
# PYSEC-2022-42969: same as above (alias)
ignore-vulns = ["GHSA-4xh5-x5gv-qwph", "PYSEC-2022-42969"]
